// apps/tourism/platform/src/types/rajatrain.ts
export interface definitions {
// ...
  rajaTrainTrip: {
    id?: string;
    trainInfo?: definitions["rajaTrainInfo"];
    tickets?: definitions["rajaTrainTicket"][];
    wantCompartment?: boolean;
    compartmentType?: definitions["rajaCompartmentGenderType"];
  };
// ...
}